requirements.txt
requests>=2.0
beautifulsoup4>=4.0
pillow>=9.0
pyinstaller>=5.0
